Add files via upload
Update: 

Still working on passing push.exercises the ID's of the exercises as opposed to the exercise name as a string.

.values_list('pk', flat=True) returns what we need. 

Trying to add exercises with "Chest" OR "Tricep" primary tags is raising a syntax error. Focusing on just "Chest" to nail down the ID issue.

I now get this error after this command:

>>> push.exercises.add(Exercise.objects.filter(primary_tag="Chest").values_list('pk', flat=True))
...
TypeError: Field 'id' expected a number but got <QuerySet [9, 10, 11, 12]>.

So, closer! Now it seems the issue is moreso that I'm passing a QuerySet as opposed to individual entries...